Initial refactor of jambonz-ui package library
