+++ jambonz-ui/rollup.config.js
@@ -1,0 +1,15 @@
+import { babel } from '@rollup/plugin-babel';
+
+export default {
+  input: './src/index.js',
+  output: {
+    file: 'build/index.js',
+    format: 'cjs',
+  },
+  external: ['react'],
+  plugins: [
+    babel({
+      exclude: 'node_modules/**',
+    }),
+  ],
+};
