+++ jambonz-ui/src/utils.js
@@ -1,0 +1,5 @@
+export const classNames = (o) => {
+  return Object.keys(o)
+    .filter(k => (o[k] === true))
+    .join(' ');
+};
