+++ jambonz-ui/styles/_buttons.scss
@@ -4,6 +4,8 @@
   padding: 23px 64px 21px;
   background-color: $jambonz;
   color: $white;
+  border: none;
+  cursor: pointer;
   text-decoration: none;
   border-radius: 30px;
   display: inline-block;
