+++ jambonz-ui/styles/index.scss
@@ -1,37 +1,35 @@
 /******************************************************************************
- * @jambonz/ui
-*******************************************************************************/
-@import '../../node_modules/@jambonz/ui/styles/vars';
-@import '../../node_modules/@jambonz/ui/styles/mixins';
-
-/******************************************************************************
- * Local Vars / Mixins
+ * Variables / Mixins
 *******************************************************************************/
 @import 'vars.scss';
 @import 'mixins.scss';
 
 /******************************************************************************
- * Layout
+ * Reset
 *******************************************************************************/
-@import 'layout.scss';
-@import 'text-layout.scss';
-@import 'markdown.scss';
+@import 'reset.scss';
 
 /******************************************************************************
- * Navi
+ * Fonts
 *******************************************************************************/
-@import 'navi.scss';
+@import 'fonts.scss';
 
 /******************************************************************************
- * Footer
+ * Typography
 *******************************************************************************/
-@import 'footer.scss';
+@import 'typography.scss';
 
 /******************************************************************************
- * Pages
+ * Buttons
 *******************************************************************************/
-@import 'pages/home.scss';
-@import 'pages/why.scss';
-@import 'pages/pricing.scss';
-@import 'pages/jambonz-ui.scss';
-@import 'pages/regions.scss'
+@import 'buttons.scss';
+
+/******************************************************************************
+ * Icons
+*******************************************************************************/
+@import 'icons.scss';
+
+/******************************************************************************
+ * Backgrounds
+*******************************************************************************/
+@import 'backgrounds.scss';
