+++ jambonz-ui/test/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html dir="ltr" lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="description" content="A testing page for raw React app usage of @jambonz/ui" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>@jambonz/ui test</title>
+  </head>
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script src="/dist/test.js"></script>
+  </body>
+</html>
