+++ pages/_app.js
@@ -1,5 +1,6 @@
+import '@jambonz/ui/css/styles.css';
 import '../src/styles/global.scss';
 
-export default function App({Component, pageProps}) {
+export default function App({ Component, pageProps }) {
   return <Component {...pageProps} />;
 }
