+++ pages/index.js
@@ -1,10 +1,13 @@
-import { nanoid } from 'nanoid';
-import classNames from 'classnames';
-import { useState, useEffect, useRef } from 'react';
+import Link from 'next/link';
 
-import Layout from '../src/components/layout';
-import { Latest, Hero, Icon, Button, H4, H3, H2, P, MS, normalizeSubtext, normalizeSlug, useMobileMedia } from '../src/components/jambonz-ui';
+import { nanoid } from 'nanoid';
+import { useState, useEffect, useRef } from 'react';
+import { Icon, Button, H4, H3, H2, P, MS, classNames, IconGroup } from '@jambonz/ui';
+
 import { getData } from '../src/lib/data';
+import { useMobileMedia } from '../src/components/hooks';
+import Layout, { Latest, Hero } from '../src/components/layout';
+import { normalizeSubtext, normalizeSlug } from '../src/components/utils';
 
 function Tech({data}) {
   return (
@@ -133,11 +136,11 @@ function BYO({data}) {
         <div className="byo__subtext">
           <H3 className="h5">{data.subtext}</H3>
         </div>
-        <div className="byo__icons icons">
+        <IconGroup className="byo__icons">
           {data.icons.map((icon) => {
             return <Icon key={nanoid()} name={icon} mainStyle="fill" />;
           })}
-        </div>
+        </IconGroup>
         <div className="byo__comment">
           <H4 className="h5">
             {/* Use dangerouslySetInnerHTML to render inline link from YAML data */}
@@ -145,7 +148,7 @@ function BYO({data}) {
           </H4>
         </div>
         <div className="byo__cta">
-          <Button href={data.url} subStyle="dark">{data.cta}</Button>
+          <Button Link={Link} href={data.url} subStyle="dark">{data.cta}</Button>
         </div>
       </div>
     </section>
