+++ src/components/navi.js
@@ -1,10 +1,11 @@
-import { useState } from 'react';
-import { nanoid } from 'nanoid';
-import classNames from 'classnames';
-
 import Link from 'next/link';
 
-import { Button, Icon, useMobileMedia, useActiveNavi, rSlash } from './jambonz-ui';
+import { nanoid } from 'nanoid';
+import { useState } from 'react';
+import { Button, Icon, classNames } from '@jambonz/ui';
+
+import { rSlash } from './utils';
+import { useMobileMedia, useActiveNavi } from './hooks';
 
 function NaviItem({obj}) {
   const activeNavi = useActiveNavi();
@@ -37,7 +38,7 @@ function NaviMobile({ active, handler, siteData }) {
     <div className={classNames(classes)}>
       <div className="navi__mobile__head">
         <div className="navi__mobile__login">
-          <Button href={siteData.navi.login.link} mainStyle="login" subStyle="white" onClick={handler}>
+          <Button Link={Link} href={siteData.navi.login.link} mainStyle="login" subStyle="white" onClick={handler}>
             {siteData.navi.login.label}
           </Button>
         </div>
@@ -57,7 +58,7 @@ function NaviMobile({ active, handler, siteData }) {
         })}
       </ul>
       <div className="navi__mobile__support">
-        <Button href={`mailto:${siteData.footer.email}`} target="_blank" subStyle="light">
+        <Button Link={Link} href={`mailto:${siteData.footer.email}`} target="_blank" subStyle="light">
           {siteData.footer.email}
         </Button>
       </div>
@@ -100,7 +101,7 @@ export default function Navi({ siteData }) {
           <Icon mainStyle="fill" name="Menu" />
         </div>
         <div className="navi__login">
-          <Button href={siteData.navi.login.link} mainStyle="login">
+          <Button Link={Link} href={siteData.navi.login.link} mainStyle="login">
             {siteData.navi.login.label}
           </Button>
         </div>
